Add JupyterLab to the upgrade playbook
I started to use JupyterLab Desktop and since there is no RPM
repository, I need to upgrade it directly from the RPM available on
GitHub.  The JupyterLab Desktop is an Electron app, and it has conflict
with Slack.  To workaround this issue, I need to upgrade both Slack and
JupyterLab Desktop via

    rpm -iU --force <package>.rpm

Luckily, the rpm command can do it directly from the URL, so there is no
need to download it in advance.