Fix range of years explanation
I prefer using range of years and to be on the safe side, I need to add
explicit explanation of meaning.  I found an existing text phrased by
someone else that seems to be legit, so I attached it to both LICENSE
and README.md.  I also fixed some minor errors in README.md.