Fix Zotero upgrade
Upgrade playbook got broken with Zotero 7.  The path of application.ini
changed, JavaScript of the Zotero Download webpage got changed.

Furthermore, during debugging process I found out that the
ansible.builtin.file module does not support glob patterns, so the only
way how to delete the contents of a directory is to delete the directory
and then recreate it in the next step.  I will need to fix this bug for
other tasks as well in the following commit.