Fix ansible.builtin.file calls
I have learned recently that the ansible.builtin.file module does not
support glob patterns.  It means that the only way how to ensure that
the directory exists and it is empty is to first delete the directory if
exists and recreate it afterwards.  My previous misconception caused
several bugs which went unnoticed up to now.